Optimizes the `ReloadableJava17TypeMapping` implementation
Reduces the number of object allocations in `ReloadableJava17TypeSignatureBuilder` by using a single `StringBuilder` when constructing complex signatures. This comes at the cost of higher complexity, so some performance measurements should be done to establish whether it is really worth it.

Additionally, the `ReloadableJava17TypeMapping` (which uses `ReloadableJava17TypeSignatureBuilder`) will now only compute the cache key (a Snappy-compression of the type signature) once on a cache miss.

Test LZ4 instead of Snappy.